Add support for sections in the URL list
URLs can be grouped together in a section in the /list subpage. Each section is sorted individually. The section levels are ignored completely (but preserved).
In the non-subpage, any list section can be used with "<!-- bot:Title -->" where "Title" is the section name (case-sensitive).
URLs appearing outside of any section (i.e. before the first section or when there are no sections) can be used as before with "<!-- bot -->".