Add a notes column
Usage: on the /list page, you can have an entry like 'https://example.org/ | note = Something' to attach the note "Something" to that entry. This ends up in a "Notes" column on the non-list page. The column is only present when at least one entry in a table has a note.

To make this more flexible for potentially adding further options in the future, I also added an explicit 'label' option name. The list can still be entered as before but will be normalised to 'URL | label = LABEL' by the bot. ('URL | LABEL | NOTE' is also valid and will be normalised accordingly.)